% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols}
\alias{fastols}
\title{Function that performs linear regression
and saves the diagonal of the hat matrix}
\usage{
fastols(X, y, w, compute_se = 1L, compute_hat = 0L)
}
\arguments{
\item{X}{an nxk numeric data matrix}

\item{y}{The nx1 numeric output vector}

\item{w}{an nx1 numeric vector of weights}

\item{compute_se}{binary; 1 to compute VCV matrix, 0 otherwise}

\item{compute_hat}{binary; 1 to compute diagonal of hat matrix, 0 otherwise}
}
\description{
Function that performs linear regression
and saves the diagonal of the hat matrix
}
