% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols_by}
\alias{fastols_by}
\title{Function that performs linear regression
and saves the diagonal of the hat matrix}
\usage{
fastols_by(X, y, w, g)
}
\arguments{
\item{X}{an nxk numeric data matrix}

\item{y}{The nx1 numeric output vector}

\item{w}{an nx1 numeric vector of weights}

\item{g}{an nx1 sorted integer vector of groups}
}
\description{
Function that performs linear regression
and saves the diagonal of the hat matrix
}
