% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols_bywt}
\alias{fastols_bywt}
\title{Function that performs linear regression
and saves the diagonal of the hat matrix}
\usage{
fastols_bywt(X, y, w, g, nthr = 1L, compute_se = 1L, compute_hat = 0L)
}
\arguments{
\item{X}{an nxk numeric data matrix}

\item{y}{The nx1 numeric output vector}

\item{w}{an nx1 numeric vector of weights}

\item{g}{an nx1 sorted integer vector of groups}

\item{nthr}{integer; number of threads to use for parallel processing}

\item{compute_se}{binary; 1 if se should be computed, 0 if not. 1 by default}

\item{compute_hat}{binary; 1 if diagonal of at matrix should be calculated,
0 if not. 0 by default.}
}
\description{
Function that performs linear regression
and saves the diagonal of the hat matrix
}
