% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_by}
\alias{loclin_by}
\title{Function that performs local linear regression
using Gaussian or Epanechnikov kernel.}
\usage{
loclin_by(X, H, y, Xeval, sameX, kernel, nthr = 1L)
}
\arguments{
\item{X}{an nxk data matrix}

\item{H}{a kxk positive definite bandwidth matrix}

\item{y}{The nx1 output vector}

\item{Xeval}{an mxp matrix at which to predict using local linear regression}

\item{sameX}{binary; Are the evaluation points the same as X? One for yes.}

\item{kernel}{integer; 1 for Gaussian, 2 for Epanechnikov}

\item{nthr}{integer; number of threads to use for parallel processing}
}
\description{
Function that performs local linear regression
using Gaussian or Epanechnikov kernel.
}
