% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_sameX}
\alias{loclin_sameX}
\title{Function that performs local linear regression
using Gaussian or Epanechnikov kernel.}
\usage{
loclin_sameX(X, y, H, kernel)
}
\arguments{
\item{X}{an nxk data matrix}

\item{y}{The nx1 output vector}

\item{H}{a kxk positive definite bandwidth matrix}

\item{kernel}{integer; 1 for Gaussian, 2 for Epanechnikov}
}
\description{
Function that performs local linear regression
using Gaussian or Epanechnikov kernel.
}
