% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_sameX_unif}
\alias{loclin_sameX_unif}
\title{Function that performs local linear regression
using Gaussian or Epanechnikov kernel.}
\usage{
loclin_sameX_unif(X, y, H)
}
\arguments{
\item{X}{an nxk data matrix}

\item{y}{The nx1 output vector}

\item{H}{a kxk positive definite bandwidth matrix}
}
\description{
Function that performs local linear regression
using Gaussian or Epanechnikov kernel.
}
