% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_sameX_unif_by}
\alias{loclin_sameX_unif_by}
\title{Function that performs local linear regression
using uniform kernel, by group}
\usage{
loclin_sameX_unif_by(X, y, g, H, nthr = 1L)
}
\arguments{
\item{X}{an nxk data matrix}

\item{y}{The nx1 output vector}

\item{g}{nx1 integer vector}

\item{H}{a kxk positive definite bandwidth matrix}

\item{nthr}{positive integer; number of threads}
}
\description{
Function that performs local linear regression
using uniform kernel, by group
}
