% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols}
\alias{fastols}
\title{Function that performs linear regression
and saves the diagonal of the hat matrix}
\usage{
fastols(X, y, w)
}
\arguments{
\item{\code{X}:}{an nxk data matrix}

\item{\code{y}:}{The nx1 output vector}

\item{\code{w}:}{an nx1 vector of weights}
}
\description{
Function that performs linear regression
and saves the diagonal of the hat matrix
}
