% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastols}
\alias{fastols}
\title{Function that performs linear regression
and saves the diagonal of the hat matrix}
\usage{
fastols(X, y, w)
}
\arguments{
\item{X}{an nxk numeric data matrix}

\item{y}{The nx1 numeric output vector}

\item{w}{an nx1 numeric vector of weights}
}
\description{
Function that performs linear regression
and saves the diagonal of the hat matrix
}
