% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linsmooth.R
\name{linsmooth}
\alias{linsmooth}
\title{This function performs linear smoothing}
\usage{
linsmooth(
  X,
  y,
  Xeval = NULL,
  method = "ols",
  w = NULL,
  kernel = "gauss",
  H = NULL,
  by = NULL,
  compute_se = TRUE,
  compute_hat = FALSE,
  nthr = 1L
)
}
\arguments{
\item{X}{numeric matrix; an nxk data matrix of observations}

\item{y}{numeric vector;an nx1 response vector}

\item{Xeval}{numeric matrix; mxk data matrix where the linear smoother should
produced predicted values of y. Defaults to X.}

\item{method}{character; "ols" for linear regression or "loclin" for
local linear regression. Defaults to "ols"}

\item{w}{numeric vector; nx1 observation weights,
not used for method "loclin", optional for "ols"}

\item{kernel}{character; kernel used if method is "loclin". Defaults to
"gauss", the Gaussian kernel.}

\item{H}{positive vector; Bandwidth vector that represents diagonal of
bandwidth matrix used if "loclin" method is used. Defaults
to a diagonal matrix being optimized by minimizing LOOCV score.}

\item{by}{an nx1 vector of groups if performing linear smoothing within
groups.}

\item{compute_se}{logical; If TRUE then OLS procedure will compute standard
errors. TRUE by default.}

\item{compute_hat}{logical; If TRUE then diagonal of hat matrix (leverage) is
calculated for each observation. FALSE by default.}

\item{nthr}{positive integer; Number of threads to use for parallelized
operations. Defaults to 1 thread.}
}
\description{
This function performs linear smoothing
}
