% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linsmooth.R
\name{linsmooth}
\alias{linsmooth}
\title{This function performs linear smoothing}
\usage{
linsmooth(X = NULL, y = NULL, method = "loclin", w = NULL, kernel = "gauss")
}
\arguments{
\item{X}{numeric matrix; an nxk data matrix}

\item{y}{numeric vector;an nx1 response vector}

\item{method}{character; "ols" for linear regression or "loclin" for
local linear regression}

\item{w}{numeric vector; nx1 observation weights,
not used for method "loclin", optional for "ols"}

\item{kernel}{character; kernel used if method is "loclin". Defaults to
"gauss", the Gaussian kernel.}
}
\description{
This function performs linear smoothing
}
