% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_diffX}
\alias{loclin_diffX}
\title{Function that performs local linear regression
using Gaussian or Epanechnikov kernel.}
\usage{
loclin_diffX(X, y, H, Xeval, kernel, nthr = 1L)
}
\arguments{
\item{X}{an nxk data matrix}

\item{y}{The nx1 output vector}

\item{H}{a kxk positive definite bandwidth matrix}

\item{Xeval}{an mxp matrix at which to predict using local linear regression}

\item{kernel}{integer; 1 for Gaussian, 2 for Epanechnikov}

\item{nthr}{positive integer; number of threads}
}
\description{
Function that performs local linear regression
using Gaussian or Epanechnikov kernel.
}
