% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loclin_gauss}
\alias{loclin_gauss}
\title{Function that performs local linear regression
using a Gaussian Kernel.}
\usage{
loclin_gauss(X, H, y)
}
\arguments{
\item{\code{X}:}{an nxk data matrix}

\item{\code{H}:}{a kxk positive definite bandwidth matrix}

\item{\code{y}:}{The nx1 output vector}
}
\description{
Function that performs local linear regression
using a Gaussian Kernel.
}
