% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loocv_gauss}
\alias{loocv_gauss}
\title{Function that returns LOOCV score using Gaussian kernel}
\usage{
loocv_gauss(X, H, y)
}
\arguments{
\item{X:}{an nxk data matrix}

\item{H:}{a kxk positive definite bandwidth matrix}

\item{y:}{The nx1 output vector}
}
\description{
Function that returns LOOCV score using Gaussian kernel
}
